backend/requirements.txt
Django==3.0.3
djangorestframework==3.10.1
django-cors-headers==3.2.1
mysql-connector-python==8.0.19
pandas